requirements.txt
APScheduler==3.10.1
asgiref==3.6.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.1.8
django-cors-headers==3.14.0
django-filter==23.1
django-rest-registration==0.8.0
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.3.0
idna==3.4
pymongo==3.12.3
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
requests==2.28.2
six==1.16.0
sqlparse==0.2.4
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
